Update core.py
There was a bug ~L235 because Tracking requires data from both the colour and monochrome cameras. If one of those was missing, could not call Tracking and startup crashed.

An alternative to what I've done here would be to create dummy cameras to substitute for the missing camera(s) (see code below).

I've tested both alternatives and the only differences I can see occur if there is a colour, but no monochrome camera detected on startup: 1) with dummy cameras, we still trigger the flashes but with the present version, the colour cameras takes photos but cannot trigger the flashes; 2) if there is a dummy camera, we get a constant printout of "Dummy camera. Real camera (probably) not found" or similar.

I've chosen to go with the simpler method because it was simpler and because I didn't want the monochrome boxes to constantly print out the dummy camera message. HOWEVER, this means that if the mono camera ever fails on the colour boxes, so will the flash triggering. If we think that's a risk, we can replace L204-246 with the following:

    cam_ids = getcameraids()
    
    for cam_id in cam_ids:
        camera = Camera(message_queue,trigger.record,cam_trigger,cam_id=cam_id)
        cameras.append(camera)
        t = Process(target=camera.worker)
        t.start()
        import time
        time.sleep(1)

    #we'll make the tracking camera the first greyscale one if there is one, ot>
    greyscalecam = None
    colourcam = None
    
    print("looking for camera to use for tracking...")
    for cam in cameras:
        print("Cam...")
        print(cam.colour_camera.value)
        if cam.colour_camera.value==0:
            greyscalecam = cam
        else:
            colourcam = cam

    if greyscalecam is None:
        print("No Monochrome Camera Found: Adding dummy camera")
        camera = Dummy_Camera(message_queue,trigger.record,cam_trigger,cam_id=l>
        cameras.append(camera)
        t = Process(target=camera.worker)
        t.start()
        import time
        time.sleep(1)
        greyscalecam=camera
    if colourcam is None:
        print("No Monochrome Camera Found: Adding dummy camera")
        camera = Dummy_Camera(message_queue,trigger.record,cam_trigger,cam_id=l>
        cameras.append(camera)
        t = Process(target=camera.worker)
        t.start()
        import time
        time.sleep(1)
        colourcam=camera

    tracking = Tracking(message_queue,greyscalecam.photo_queue,colourcam.photo_>
    t = Process(target=tracking.worker)
    t.start()