Avoid crash at L30 if object doesnt have self.index
I've had several crashes when a tracking object gets to L30 and crashes because it has no self.index attribute to print. This quick fix stops that crash but doesn't tackle the underlying problem: presumably Mike never expected tracking objects to set their attributes through this route and some setting has failed to be applied because of this issue.

I only encountered this issue for the firsts time in Ohio in May 2025, so presumably a recent change to the code has caused it, but I don't know what.